requirements.txt
dash_bootstrap_templates
Flask-Caching
pymssql
pandas
SQLAlchemy
